Carousel completed and folder restructured
