+++ src/App.js
@@ -1,6 +1,7 @@
-import React from 'react';
-import './App.css';
-import Home from './components/Home Page/home/home';
+import React from "react";
+import "./App.css";
+import Home from "./containers/Home/home";
+// import Login from "./containers/Login/login";
 
 function App() {
   return <Home />;
