--- src/components/Home Page/carousel/carousel.css
@@ -1,1 +1,0 @@
-/* .carousel-css {} */
