+++ src/components/card/card.css
@@ -1,8 +1,11 @@
 .card {
-    box-shadow   : 0 4px 8px 0 rgba(0, 0, 0, 0.2);
-    transition   : 0.3s;
-    width        : 1300px;
-    border-radius: 5px;
+    box-shadow    : 0 4px 4px 0 rgba(0, 0, 0, 0.2);
+    transition    : 0.3s;
+    width         : 300px;
+    border-radius : 5px;
+    display       : inline-block;
+    margin        : 0 10px 16px 10px;
+    padding-bottom: 5px;
 }
 
 .card:hover {
@@ -18,7 +21,7 @@
 }
 
 .button-css {
-    margin-left     : 23%;
+    margin-left     : 32%;
     height          : 30px;
     background-color: #2d0ce4;
     color           : white;
