+++ src/components/toast/toast.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import './toast.css';
+
+const ToastMessage = ({ message }) => {
+  return <div className="toasted">{message}</div>;
+};
+
+export default ToastMessage;
